# backendapi/src/main/resources/openapi.yaml
openapi: 3.0.0

info:
  title: API de Usuarios

  description: API para gestionar usuarios (CRUD)

  version: 1.0.0

servers:
  - url: http://localhost:8080/api

    description: Servidor de desarrollo

paths:
  /usuarios:
    get:
      summary: Obtener todos los usuarios

      responses:
        "200":
          description: Lista de usuarios

          content:
            application/json:
              schema:
                type: array

                items:
                  $ref: "#/components/schemas/Usuario"

    post:
      summary: Crear un nuevo usuario

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"

      responses:
        "201":
          description: Usuario creado

  /usuarios/{id}:
    get:
      summary: Obtener un usuario por ID

      parameters:
        - name: id

          in: path

          required: true

          schema:
            type: integer

      responses:
        "200":
          description: Usuario encontrado

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"

    put:
      summary: Actualizar un usuario por ID

      parameters:
        - name: id

          in: path

          required: true

          schema:
            type: integer

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Usuario"

      responses:
        "200":
          description: Usuario actualizado

    delete:
      summary: Eliminar un usuario por ID

      parameters:
        - name: id

          in: path

          required: true

          schema:
            type: integer

      responses:
        "204":
          description: Usuario eliminado

components:
  schemas:
    Usuario:
      type: object

      properties:
        id:
          type: integer

          format: int64

          description: ID del usuario

        nombres:
          type: string

          description: Nombres del usuario

        apellidos:
          type: string

          description: Apellidos del usuario

        rut:
          type: integer

          format: int64

          description: RUT del usuario (sin dígito verificador)

        dv:
          type: string

          description: Dígito verificador del RUT

        fechanacimiento:
          type: string

          format: date

          description: Fecha de nacimiento del usuario (YYYY-MM-DD)

        correoelectronico:
          type: string

          description: Correo electrónico del usuario

        contraseña:
          type: string

          description: Contraseña del usuario

      required:
        - nombres

        - apellidos

        - rut

        - dv

        - fechanacimiento

        - correoelectronico

        - contraseña